refactor(desktop): group custom protocols into main/protocols
