+++ apps/desktop/src/main/protocols/art-protocol.ts
@@ -2,11 +2,11 @@ import { app, protocol, nativeImage } from 'electron';
 import * as fs from 'fs';
 import * as path from 'path';
 import * as crypto from 'crypto';
-import { logger } from './app/logger';
+import { logger } from '../app/logger';
 import { tracks, playlists } from '@shiranami/database';
 import { getDatabase } from '@shiranami/database/client';
-import { artUrlFor } from './shared/album-art-image';
-import { IMAGE_EXTENSIONS, imageMime } from './shared/media-types';
+import { artUrlFor } from '../shared/album-art-image';
+import { IMAGE_EXTENSIONS, imageMime } from '../shared/media-types';
 
 /** Directory where extracted album art images are stored */
 let artDir: string;
