+++ apps/desktop/src/main/protocols/privileged-schemes.ts
