+++ apps/desktop/src/main/protocols/protocol-registration.test.ts
