+++ apps/desktop/src/main/services/metadata-enrich-batch.ts
@@ -133,7 +133,7 @@ export async function enrichSingleTrack(
     // Either preview mode or apply-without-file-write: cache the cover so the
     // renderer can display + commit it via DB-only update. Orphaned cache
     // entries (preview-then-discard) are harmless and dedupe by content hash.
-    const { saveAlbumArt } = await import('../art-protocol');
+    const { saveAlbumArt } = await import('../protocols/art-protocol');
     const albumArtUrl = await saveAlbumArt(coverImageBuffer, coverImageMime);
     if (albumArtUrl) {
       updatedFields.albumArt = albumArtUrl;
