+++ apps/desktop/src/main/services/metadata-service.ts
@@ -2,7 +2,7 @@ import * as path from 'path';
 import type { TrackMetadata } from '@shiranami/contracts';
 import { UNKNOWN_ARTIST, UNKNOWN_ALBUM } from '@shiranami/shared';
 import { logger } from '../app/logger';
-import { saveAlbumArt } from '../art-protocol';
+import { saveAlbumArt } from '../protocols/art-protocol';
 import { isAudioExtension } from '../shared/media-types';
 
 export type { TrackMetadata };
